* rename Anvil.load to Anvil.loadRaw and create a new Anvil.load that does the nbt -> prismarine-chunk conversion * expose things properly * delete example.js * adapt example_read * remove dependencies to minecraft-mca and minecraft-region
