+++ example_write.js
@@ -1,4 +1,4 @@
-var Anvil=require("./dist/anvil");
+var Anvil=require("./").Anvil;
 var Vec3 = require("vec3");
 
 var chunk = new Anvil(process.argv[2] ? process.argv[2] : "world/lttp");
@@ -6,8 +6,8 @@ var chunk = new Anvil(process.argv[2] ? process.argv[2] : "world/lttp");
 var c=chunk.save(-32,0,{"name":"","type":"compound","value":{}});
 
 c.then(function(){
-    console.log("saved");
-  })
-  .catch(function(err){
-    console.log(err.stack);
-  });
+  console.log("saved");
+})
+.catch(function(err){
+  console.log(err.stack);
+});
