+++ package.json
@@ -36,8 +36,6 @@
   "dependencies": {
     "babel-runtime": "^5.4.4",
     "es6-promisify": "^3.0.0",
-    "minecraft-mca": "^0.6.1",
-    "minecraft-region": "rom1504/minecraft-region#patch-1",
     "node-promise-es6": "^0.2.3",
     "prismarine-chunk": "^0.2.1",
     "prismarine-nbt": "^0.1.0",
