+++ src/index.js
@@ -1,0 +1,4 @@
+module.exports={
+  Anvil:require("./anvil"),
+  chunk:require("./chunk")
+};
